+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### New Features
 - Added ``--load-checkpoints`` to load particular checkpoints
+- Added ``--num-threads`` to enjoy script
 
 ### Bug fixes
 - Fixed optuna warning (@SammyRamone)